pyproject.toml
[project]
name = "FoldSeekStrucAnnoFlow"
version = "0.0.1"
description = "The configuration file for any python scripts used in the FoldSeek structural annotation Nextflow pipeline."
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [{ name = "Luc Elliott", email = "luc1997@liverpool.ac.uk" }]
dependencies = [
  "biopython",
  "click",
  "typer",
]

[dependency-groups]
dev = [
  "black",
  "build",
  "coverage",
  "isort",
  "mypy",
  "pydantic",
  "pytest",
  "pytest-cov",
  "pytest-pretty",
  "ruff",
  "tombi",
  "uv",
]

[build-system]
requires = ["setuptools>=65.5.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.coverage.run]
omit = [
  "tests/*",
  "noxfile.py",
  # add other files or directories to omit from coverage here
]

[tool.mypy]
check_untyped_defs = true
disallow_incomplete_defs = true
disallow_untyped_calls = false
disallow_untyped_defs = true
no_implicit_optional = true
plugins = ["pydantic.mypy"]
strict_optional = true
warn_return_any = true
warn_unused_ignores = true
exclude = [
  ".nox",
  "noxfile.py",
  "tests/*",
  "FoldSeekStrucAnnoFlow/external/*",
]

[tool.ruff]
line-length = 120
exclude = [
  ".venv",
  "README.md",
  # add other files or directories to exclude from linting here
]

[tool.setuptools.packages.find]
include = ["foldseekstrucannoflow*"]
exclude = ["works/*"]
